requirements.txt
Django==2.1.7
django-simple-captcha==0.5.9
celery==4.2.2
django-celery-results
requests-html
eventlet
redis